feat: move useDebounce hook to a separate file for better reusability
